moved customization components into seperate folder
