+++ client/src/components/characterCustomization/Avatars.jsx
